refactor: move all env factories to a single env struct (#28)
Each dimension which commands might require is a different trait, and commands
should require a type that matches the trait it needs. So we can now create
tests for commands executors, and inject a special environment to mock
whats needed.

This is a very crude version of IoC. It works very well for us as everything
is top down (main is responsible for creating the env object that is passed
around).