refactor: move the command descriptions to YAML
This makes it easier to collaborate with the tech writing team. It
also makes it easier to script with (e.g. when we have checks for
backward compatibility).

Although we technically leak memory, we really do it only once per
yaml. Every other calls to that code will reuse a shared box.