feat: dfx start now starts an http server frontend
It will respect address and port from the config (which it did not
previously), and will proxy calls to /api/ to the port 8080 locally
which is always reserved by the client.

Also added tests for (some parts of) the config.

Also now waits and blocks the process. When the user Ctrl-C or kill
dfx it will kill the nodemanager and the http server.