feat: arguments and deserialization of return value
This adds serialization of (1) argument and deserialization of the
return value of a canister.

We support serialization through a --type flag that gives us the
type of the argument, and deserialization through a series of trial
which prints the first success (try deserializing as a u64, then
a string, ...). This is crude but until we have a full IDL text
format we wont be able to do better.