Add uninstall script
I was planning originally to add a generic dfx clean-cache
function. This would be more complex initially and would still require
for us to provide a delete all possible install location script. A
user might use a different location for each install. This is probably
going to cause issues, but a good cleanup/uninstall script helps in
this case.

As a follow up, a better approach would be to provide a cache cleanup
command, that can be used by the upgrade script.

SDK-555