Use -E option for sed; Ensure sed is available
We are now using an autoformatter (:+1:). However, we have some script
that incorrectly gets formatted, and needs to be reformated via
sed. Sed -r (extended regular expression option) is not POSIX/widely
supported, with the main culprit being Mac OS X. The -E option is
instead provided which is available at 4.2 GNU apparently (as an alias
of -r).

We also need to ensure we require sed prior to continueing installation.