nix/e2e-tests.nix: use niv instead of builtins.fetchGit (#150)
We should use `builtins.fetchGit` as less as possible because it
fetches at evaluation time (causing more stress on the hydra-evaluator
and making evaluation slower in general). And it's not supported in
strict evaluation mode which we want to move to at some point.

For now we should only use `builtins.fetchGit` for fetching private
repositories.

All other sources should be fetched using `niv`. This is easier anyway
because `niv` handles hash updates.