Add a jobset for doing SDK releases (#164)
ci/release.nix is called by the sdk-release jobset on hydra. It calls
release-jobset.nix which returns the dfx-release and
install-sh-release derivations but only when the git revision has been
tagged with a release tag (v0.1.2). In that case the version in those
derivations is set to the version in the tag ensuring that they get
rebuild.