Import motoko without passing nixpkgs
it seems cleaner to not pass `nixpkgs` here, so that the SDK imports
motoko built against, say, the same ocaml library versions that the
Motoko team develops and tests again.

Also, the SDK should be able to upgrade its nixpkgs dependencies without
worrying about breaking the motoko build, or having to synchronize with
the Motoko team.

This way you also benefit from the hydra cache, getting the derivations
cached from the motoko CI.

(Just a draft, because it may be needed to pass `nixpkgs.system` here,
and allowing that requires a small extension of the Motoko
`default.nix`. I can do that if this build fails.)