refactor: refactor IDL.js into TypeScript
Including (but not limited to): adding types everywhere, adding missing
types for some dependencies, add prettier and changing some lint rules,
fixing the tests too, remove code that is unused, removing babel as we
dont need it, and some other changes that are too much to list.

The tests all pass.