refactor: remove leb128 and remove some useless functions (#240)
Since we never use bn there is no use for us to actually use leb128, instead
we have a simple implementation which uses numbers.

This also changes the IDL API to using BigNumber instead of plain Javascript numbers. Since the API hasn't been released yet, this is not considered a breaking change.