feat: add basic support for frontend development
This adds a webpack configuration and additional files when creating
new projects, if either node is installed or --frontend is used. On
dfx build, if a package.json file is present we run npm run build.

There are issues with this flow (notably the lack of watch mode),
but this is the first step and I hope to refine this before 0.4.10./