ci: add the dfinity-sdk.packages.cargo-security-audit job (#265)
The `dfinity-sdk.packages.cargo-security-audit` job only exists in the
`sdk` jobset (i.e. the jobset which corresponds to the `master`
branch).

The job will run `cargo audit` which scans `Cargo.lock` for security
vulnerabilities in crates reported in the RustSec advisory-db. The DB
is an input to the jobset. This means that whenever a new
vulnerability is reported or when `Cargo.lock` changes `cargo audit`
will run.