nix: upgrade common to fix the cargo-security-audit job
Previously `cargo-security-audit` didn't produce the required $out when run
successfully. Now the $out produced is the output from `cargo audit`.

The output is also registered as a Hydra build product so that we can
directly see the report from Hydra. The latter is useful for checking
which warnings were reported.