fix: Canister ids are little endian, not big endian (#303)
* Canister ids are little endian, not big endian

The client has switched to presenting canister ids as blobs on the
System API, but not the HTTPs Interface, where it still takes u64.
This means that the decision how to encode the u64 is not purely
internal to the client, and as long as the SDK works around that
interface, the SDK needs to make the same decision as the client.

In the System API, the client uses little endian encoding. So we should
use that here too.

With this change, I can paste the `ic:…` id for an installed canister
into some motoko source (`actor "ic:…"`) and it works. (Before this
change I had to manually reverse the bytes).

As soon as the the client updates the interface to _actually_ expect a
blob as the canister id, this will disappear from the SDK, and become
properly internal.

* cargo-fmt

Co-authored-by: chenyan-dfinity <48968912+chenyan-dfinity@users.noreply.github.com>