Userlib: CaniterIds are little endian
this is the equivalent of what #303 did to `dfx`: When CBOR-encoding a
canisterid for submission in a request, we need to make sure it is
little endian.

We do the change in
```
class CanisterIdEncoder implements CborEncoder<CanisterId> { …
```
where it belongs: As close to the HTTP interface we can.

This means that once we migrate to
https://github.com/dfinity-lab/dfinity/pull/2286
we will adjust this code to switch to CBOR blob, and the work-around can
be dropped here. Note in the code about that.

Also cleans up `fromText` to be more robust: It will _only_ accept `ic:`
textual representation, and _always_ strip the checksum.

With this, I can
 * build and install Linkedup
 * copy the `ic:` url printed by `dfx` upon installation
 * use that in the url (e.g. `http://127.0.0.1:8000/?canisterId=ic:72028DEDFB546FA89A`)
 * see it load the assets