feat(ic_http_agent): add first draft of Agent struct (#279)
* feat(ic_http_agent): add first draft of Agent struct

It will replace the ApiClient we have, but also can be published as a
crate for Rust projects to communicate with the client. It understands
Blobs as well as IDL queries right now.

A follow up PR will add call, install and request status, as well
as using it in dfx.