public: rename "unreleased" back to "latest" & always include a rev
To quote: https://dfinity.slack.com/archives/CL7Q2RXUM/p1582295540092400

It seems that we serve the latest installer script from the `sdk`
`master` branch. Shouldn't we serve the latest *released* installer?

This is how I understand things are working now:

* https://sdk-ext.dfinity.systems/install.sh redirects to: https://sdk-ext.dfinity.systems/downloads/public/latest/installer
* https://sdk-ext.dfinity.systems/downloads/public/ proxies to the blobules S3 bucket at the path `sdk/public`
* So https://sdk-ext.dfinity.systems/downloads/public/latest/installer fetches https://blobules.dfinity.systems/sdk/public/latest/installer

The problematic bit is that `latest` got renamed to `unreleased` so
we're now serving a slightly out of date installer.

This commit renamed `unreleased` back to `latest`. However, we should
discuss why were not serving the latest *released* installer.

This commit also now always includes the `revision` in the install
script regardless of whether we're doing a release or not.