feat: add verbose and quiet flags
The flag is currently exclusively for the root application, so it's dfx -v canister call ... and not dfx canister call -v .... Same for -q.

Verbose and Quiet are cumulative and subtracted from each others. So dfx -q -q -v -v will be the same as if there was no flags at all.

Having a negative count of quiet flags will also disable progress bars and spinners.