feat: Integrate identity in agent and dfx
We introduce a signer trait that provides principals and signatures
for every submit request. Right now signatures are only expected for
submit requests by the HTTP handler. Furthermore, there is a weird
limbo regarding the envelope type that needs clarification, as to when
it is going to be supported.

About Signer:
For the agent only signatures matter at the end of the day. Thus
signer provides a sign trait that takes a serializable type and
provides a request id with a serializable value containing a
signature. Signer is not parameterized on the type on purpose.

Recall, that currently object safety rules instruct that parametric
functions can not be members of trait objects. There is a way to
bypass that rule, by instantiating a container object that implements
both the resulting trait and realizes a trait monomorphizes on the fly
the original function.

However, Serialize is also one such trait. To keep this sane and
simple, and not parameterize the trait itself, we accept and return
Serialize types in the associated sign() function.

Regardless we end up with a complicated valie validity story though,
as we need to ensure ownership. Note we pass values with elided
lifetimes. That is currently not good enough for async traits
though. To avoid polluting with lifetime parameters, and in search of
a better path, we keep sign as a non-async function for now.