feat: Add identity stage1 signing (#462)
* feat: Add identity manager

* address; change to non-ephemeral

* Document; minor

* Rv partial changes

* Add signer trait to the agent

* rename

* fixup! rename

* fixup! rename (2)

* Add AsRef for Principals

* Document; Add Readme; Rm main; Rename module to types

Removed main as that is no longer necessary for development.

* Move improvements to skeleton branch

Checks if the provided path contains a creds.pem file. Creates it as
necessary. Note this is temporary, as we care only about a single key
pair right now.

* Remove interfaces for now

* Integrate identity manager attempt 2

* Add profile path under .cache/dfinity/

This ensures we do not have any user credentials and information under
a git repository. When the state crystalizes a user may also migrate
that to a different system, if they want to use encrypted PEM files
approach. (This is proposal as discussed on Friday.)

* Store Key files locally in a project

* Remove documentation

Will add in separate PR as a prelude to stage 2 (or 3).

* Add default subfolder

* Rm principal from identity

* Unupdate

includes rerun

* Compactify

* Comment

* Fix comment