refactor: remove blobules publishing code
Now that we publish directly to the https://download.dfinity.systems
CDN we don't need to publish to blobules anymore. So this commit
removes and cleans up any code related to that.

This commit also refactors the `install.sh` script generation and
makes sure `shellcheck` and `shfmt` are correctly applied to the
script.