refactor: move nix/mk-release.nix to publish.nix
Move `nix/mk-release.nix` to `publish.nix` since we only use `mkRelease` there. The function is also renamed to `mkDfxTarball`.

Additionally it's simplified as follows:

* Use `runCommandNoCC` instead of `stdenv.mkDerivation` which makes it
easier to read.

* Replace Nix anti-quotations by environment variables. 
This makes the build command easier to read and makes 
the nix-diff between two versions of the resulting .drv files
easier to parse.

* Don't store the tarball in a directory in `$out/` but store it
directly as a file named `$out`. This is simpler to work with.

* Put the version in the derivation name so you get a path like:
`/nix/store/dz7dkm4m7cxj5pmczwjxd7jsmgpv95vd-dfx-0.5.5.tar.gz`
This is nicer for debugging because you can directly see which version
of `dfx` is being build.