refactor: the buildInputs of node-e2e-tests
This simplifies the `buildInputs` of the `node-e2e-tests` derivation.

The `PATH` setting was uncessary since all bins in `buildInputs` are
automatically added to the `PATH` by `stdenv.mkDerivation`.

The `coreutils` dependency was also uncessary since it's included by
default in `stdenv.mkDerivation`.

The `buildInputs` of the `userlib-js` derivation had to be repeated in the `node-e2e-tests` derivation which is prone to errors. Now we propagate the inputs from `userlib-js` to `node-e2e-tests`.