refactor: move userlibs into the agent directory (#586)
* refactor: move userlibs into the agent directory

This is to officialize the rename from userlib to agent. It is expected
that the JavaScript agent will be published as @dfinity/agent, and the
Rust agent will simply be ic-agent.

The versions of those libraries should follow DFX version and release
schedule. The release instructions haven't been updated yet, and will
be part of a follow up PR.

ref #585

* fixup: fix the gitinore files that were dropped

* fixup: add js agent to package.json

* fixup: fix node-e2e bug