refactor: rework the replica_api and ic_agent public API
We need to fully hide the replica_api module as it will soon be generated
by our build script. As such we cannot export it and must export our own
types that map onto it.

Since the signature only requires the RequestId, we only pass that to the
Signer trait (which is now a more general Identity). I also added the
`sender()` method to the Identity trait as this is required before
calculating the RequestId. So the flow is now:

- Agent asks for the sender() on every requests
- Agent creates the Request object and calculates its RequestId
- Identity signs using the RequestId, also receives the Principal returned
  by the first step (as a courtesy if there are more than one Principal).
- Agent fills in the Envelope with the Request and the signature.
- Sends to server, the rest is essentially the same flow.

This means that now every requests has a sender (except request_status).

A next step will be to add APIs for specifying the identity PER REQUEST,
but since our API is the same at this point (outside of identity) I don't
want to put features in this refactor.

The rest (moving traits and dummy to a new ic_agent::identity module, adding
serializer and deserializer to Principal, adding TryFrom and From around)
are simply general Quality of Life changes.