Add field to dfx.json: defaults/build/packtool (string)
In the example of [vessel](https://github.com/kritzcreek/vessel), this would be set as follows:
```
{
...
"defaults": {
"build": {
...
"packtool": "vessel sources"
}
},
...
}
```
Or with `dfx config defaults/build/packtool "vessel sources"`

If present, this command will be called and its output passed to `moc` during both build phases of the canister pool, and to `mo-ide` for the `_language-service` command.  The package parameters are not passed when calling `moc --print-deps`.

See also https://github.com/dfinity-lab/sdk/issues/521