feat: add a custom external tool builder (#669)
This tool allows for any external process to be run and its output to
be taken and used.

This is a generalization of a Rust builder. It could also be used for
using WASM that is pre-compiled and use it without the copy wasm trick.
The build command can be missing in which case the tool will simply
use the wasm already available on disk.

This will allow us to implement new language support without specific
builders, and without divulging our plans to others. It would also
allow community members to implement their own WASM build.