feat: use local network rather than defaults.start (#766)
* feat: add network mapping in dfx.json

Add --network <network> parameter to build and canister commands.  Can be specified instead of provider.  Looks up

$ dfx build --network <network>
$ dfx canister --network network>

The ping command argument will be first looked up in dfx.json networks, and if not found, used as the URL to ping.

$ dfx ping    # uses "local" network - the default
$ dfx ping local
$ dfx ping tungsten

* Account for ic-ref ping responses

* troubleshooting: succeeds locally, of course

* Just search for ic_api_version in quotes

* Set version 1 1 in matrix-multiply dfx.json

Intended to fix warning: You seem to be running an outdated version of dfx.
and CI failure

* Try specifying bind address in dfx.json

* wip

* Remove defaults.start section

* adjust test

* format

* Fix merge error