fix: throw error on request failure and worker fix
Before, HTTP failures (like 401) were not propagated properly and
would result in decoding failure. Now at least an exception is
thrown with body and code of the failure.

Added a worker to the SiteInfo that is true if the SiteInfo was
built inside the worker (or if its from the page). This allows
getHost to know the proper host (instead of using Unknown kind)
and getWorkerHost to still return empty as we are in the worker.