refactor: refactor the whole actor and add management canister (#804)
This PR added some new types which are nice, like having an InterfaceFactory type. It also clarifies how Actor actually acts (as a class). You can still use the `makeActorFactory` method on an agent to create an actor factory, so we don't break existing code that don't use undocumented API.

The changes in the Node e2e are faulty types that were wrong before (and got uncovered by stronger typing on Actor class), and the install flow now uses the management canister and had to be changed.

Before; `install` and `createCanister` were part of an Actor instance, which didn't work so well (you need a canisterId to `createCanister` which is awkward).

Now; `install` takes a canister ID and is part of the class itself and not an instance method. There is a `createCanister` which only accepts configuration to talk to the replica. So this API is now more in line with expectations from a developer.

Because of the change to using a management canister, this puts the JS Agent to 0.8 compatibility.

Breaking changes;
- When creating an actor, you need to pass in the canister ID. This simplifies a lot of things but might break people who use the _undocumented_ API.
- People relying on methods starting with `__` will break as those are not implemented anymore.