Don't invoke ci.nix from default.nix
This moves the publish job to the `ci/` directory and make it
platform-independent. This means the jobs can now be run locally like
so:

``` shell
$ nix run -f ./ci/ci.nix publish.dfx -c activate
```

NOTE: The publish job will need to be changed; the target is now
`publish.dfx` (was `publish.dfx.x86_64-linux`). In theory we need to
make sure that machines for (1) evaluating and (2) deploying have the
same platform; in practice this is always the case.