refactor: move canister creation and installation methods to an operations module (#1003)
Moves 2 functions out of the `commands/canister` module to the `operations/canister` module.  There are no changes to the functions themselves other than the change of location.

The new files are named after the function (`create_canister.rs`, `install_canister.rs`), rather than `create.rs` or `install.rs` to make room for operations that apply to more than one canister.

This is in preparation for the next operation, to be stored in `deploy_canisters.rs`, which will have private functions and will call both of these moved functions.