refactor: when starting webserver, do not create an extra thread (#1037)
HttpServer.run() spawns worker threads.  We don't need to also start another thread in which to run it.

This simplifies the webserver start functions, which can now return the Server itself instead of sending it through a channel.

There changes proxy.rs as little as possible because it will soon be deleted due to rework of the start command.