refactor: revert: "when starting webserver, do not create an extra thread" (#1048)
Reverts dfinity-lab/sdk#1037

This unexplained thread may have a purpose: removing it triggered some sort of condition the result of which was a segmentation fault on program exit, but only in conjunction with an actor wrapper for the web server used by `dfx bootstrap`.

As it turns out, `dfx bootstrap` is already actor-based.  `dfx replica` needs an actor wrapper, because it is managing an external process, but the webserver that `dfx bootstrap` creates is already an actor behind the scenes.