fix: Avoid connection hang on Linux (#1247)
When deploying cancan on local Linux machine, dfx bootstrap server will hang during asset upload.

The problem appears to be that TCP connection to the replica's HTTP port was forever stuck in ESTAB state (according to `ss -iepn | grep tcp`). [A similar problem was fixed in reqwest](https://github.com/seanmonstar/reqwest/issues/1018) by setting keepalive. However it does not seem to work for actix-web.

I hacked through a set of libraries to set keepalive to the socket, but in the end it still hang even though I could confirm keepalive was set by using `ss`.

The fix is to limit to only 1 TCP connection per worker to connect to the replica port for now.