docs: Modify the get-wallet and set-wallet usage msgs (#1340)
Hopefully, this is the right place. Minimal changes (verb tense for consistency, remove what happens when there's no wallet because, presumably, there's a reasonable message displayed in that case so the usage doesn't need to say anything and, for most commands, we don't).

Really good write-up about wallet integration BTW. Thanks for doing that!