feat: implement the HTTP Request proposal in dfx' bootstrap webserver (#1483)
And add support for http requests in the base storage canister (with a default to `/index.html`).

This does not support other encodings than `identity` for now (and doesn't even return any headers). This support will be added to the upgraded asset storage canister built in #1482.

Added a test that uses `curl localhost` to test that the asset storage AND the webserver properly support the http requests.

This commit also upgrades tokio and reqwest in order to work correctly. There are also _some_ performance issues noted (this is slower than the `icx-http-server` for some reason), but those are not considered criticals and could be improved later on.

Renamed the `project_name` in our own generated assets to `canister_name`, for things that are generated during canister build (and not project generation).

Ref https://github.com/dfinity/dx-triage/issues/96