Internet Identity integration
this is just a proof of concept, to show the direction we can go here. I
hope someone else feels inspired to pick this up.

The subcommand names are bad, and I am unsure if they should really be
under `identity`, or rather maybe `dfx ii`?

I also wonder if `dfx` should be able to do stuff that is not related to
a developer’s project outside of a `dfx.json` file.