fix: Avoid proxy server hang on Linux by setting connection limit to 1 (#1784)
This is essentially the same fix as in #1247.

We are still seeing this bug https://forum.dfinity.org/t/multiple-calls-queries-to-canister-blocks-them-all-from-sending-back-the-response/6406.

I believe by default there are 4 worker threads so concurrent connections are still possible. Unless we have strong evidence that this kills performance, I think it is better to unblock our users before we find a permanent solution.