feat: make cargo build work, no nix required (#1887)
**TL;DR: `cargo build` now works as expected.**

The intent of this PR is to allow building `dfx` with `cargo build`, without having to install nix, by providing an alternate mechanism for preparing the assets.

- `cargo build` builds the required assets.  It stores them in`$OUT_DIR`, which means that `cargo clean` will clean them.
- `cargo build` will also rebuild the assets, iff any of their inputs changed.
- The `DFX_ASSETS` environment variable is still respected:
  - The nix build still uses it.
  - If for some reason building the assets automatically doesn't work for you, you can run `scripts/prepare-dfx-assets.sh`, and set the `DFX_ASSETS` environment variable to the directory in which you stored the assets.

It's now necessary to run `scripts/write-dfx-assets-sources.sh` when updating agent-rs, `replica`, `ic-starter`, or `ic-ref`.  The process for updating each of those is already manual.  This PR adds a CI step to make sure that this has been done.