feat: dfx deploy --mode=reinstall <canister> (#1890)
Adds a command-line argument `--mode=reinstall` to `dfx deploy`.  When present, a canister name is required.  This will reinstall the specified canister during deployment.

Because this is a destructive operation, `dfx deploy` will display a warning and prompt the user for an approval.

Also, `dfx canister install --mode=install` now requires a canister name and will fail if `--all` is passed.