chore: Update mergify configuration to account for deprecations
For:

The configuration uses the deprecated strict mode of the merge action.
A brownout is planned for the whole December 6th, 2021 day.
This option will be removed on January 10th, 2022.
For more information: https://blog.mergify.com/strict-mode-deprecation/

The configuration uses the deprecated commit_message mode of the merge action.
A brownout is planned for the whole March 21th, 2022 day.
This option will be removed on April 25th, 2022.
For more information: https://docs.mergify.com/actions/merge/