feat: support canister sandboxing (#1994)
This adjusts the scripts to download the `canister_sandboxing`
binary, which is a dependency of the `replica` binary.

Note that the current version of the replic from Dec 10 does not
have canister sandboxing flag enabled, so the downloaded
`canister_sandbox` binary is not used.

The next update of the replica binary will enable canister
sandboxing on Linux and will start using the corresponding
`canister_sandbox` binary.