feat: remote canister support (#2010)
Adds a `remote` section to canisters defined in dfx.json, which indicates that a canister already exists on certain networks and is managed by some other project.

The `remote` section can specify a canister ID per network, meaning dfx will use that canister ID when building other canisters, and will not create or install those canisters.

The `remote` section can also specify a `candid` file to use when building for remote networks, rather than the candid file either generated by `moc` or specified elsewhere for rust and custom canisters.

This facilitates writing mocks for remote canisters.  Additional public methods can be added for setting up data or responses for tests.  When building for a remote network, the `remote.candid` field is used in order to report inter-canister calls to those methods as errors.