feat: dfx deploy --upgrade-unchanged to upgrade wasm even if it did not changed (#2127)
# Description

Make it so it's possible to "upgrade" a canister's wasm even if the wasm hash did not change.  This has been asked for in order to be able to re-run the upgrade path repeatedly (even though the code didn't change).

Also applies to dfx canister install --mode=upgrade --upgrade-unchanged

# How Has This Been Tested?

Through end-to-end tests, and by running `dfx deploy --upgrade-unchanged` and `dfx install --all --mode=upgrade --upgrade-unchanged`

# Checklist:

- [x] The title of this PR complies with [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
- [x] I have edited the CHANGELOG accordingly.
- [] I have made corresponding changes to the documentation.

closes #2015