fix: report error context (#2132)
# Description

Sometimes dfx prints an error message like "No such file or directory", with no further context.

The purpose of this PR is to also display context, so that the developer has a chance of figuring out what is wrong.

Fixes https://dfinity.atlassian.net/browse/SDK-428

# How Has This Been Tested?

There are end-to-end tests that verify the new behavior.

# Checklist:

- [x] The title of this PR complies with [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
- [x] I have edited the CHANGELOG accordingly.
- [x] There were no corresponding changes to documentation.