feat: Use null as default value for opt arguments (#2153)
# Description

Before this commit, `deploy`ing a canister with an `opt Foo` init
argument without specifying an `--argument` would lead to an error:

```
$ dfx deploy
Error: Invalid data: Expected arguments but found none.
```

With this change, this isn't an error anymore, but instead `null` is
passed as a value. In general, if the user does _not_ provide an
`--argument`, and if the init method expects only `opt` arguments, then
`dfx` will supply `null` for each argument.

Note in particular that this does not try to match `opt` arguments for
heterogeneous (`opt`/non-`opt`) signatures. Note moreover that this only
impacts a case that would previously error out, so no existing (working)
workflows should be affected.

# How Has This Been Tested?

I've only manually tested this with a canister that expected a single `opt` init argument.

# Checklist:

- [x] The title of this PR complies with [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
- [x] I have edited the CHANGELOG accordingly.
- [ ] I have made corresponding changes to the documentation.