fix: Remove false error from deposit-cycles (#2155)
`dfx canister deposit-cycles` attempts to get the new balance so it can log it, and propagates any errors. This means that if it successfully deposited cycles to a canister it can't query because it doesn't own, it will visually fail with an error while the transaction proceeded just fine. This removes that behavior.