fix: install script: use either of shasum or sha256sum (#2160)
# Description

On Linux systems, `shasum` may not be available, and even if it is, it depends on perl being installed.
On OSX, `sha256sum` isn't installed by default, but can be by homebrew.

Use whichever is available, favoring sha256sum.

Also, pushd and popd aren't available in sh, so use a subshell.

# How Has This Been Tested?

Tested manually.