refactor: Environment.get_network_descriptor() returns a ref, not an option (#2220)
# Description

This eliminates a bunch of unwrap() and expect() calls that can never trigger, because the Environment in question is always an AgentEnvironment anyway.  In practice, Environment::get_network_descriptor can never be called, so it now calls `unreachable!()`.

This is preliminary work for system-wide dfx start

# How Has This Been Tested?

e2e tests will cover it

# Checklist:

- [x] The title of this PR complies with [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
- [x] (not needed) I have edited the CHANGELOG accordingly.
- [x] (not needed) I have made corresponding changes to the documentation.