feat: Allow users to configure logging level of bitcoin adapter (#2310)
The bitcoin adapter's logging can be very verbose if `debug` logging is enabled, making it difficult to make sense of what's going on. On the other hand, these logs are useful for triaging problems.

To get the best of both worlds, this commit adds support for an additional configuration option in `dfx.json`:

```
  "defaults": {
    "bitcoin": {
      "enabled": true,
      "nodes": ["127.0.0.1:18444"],
      "log_level": "info" <------- users can now configure the log level
    },
```

By default, a log level of "info" is used, which is relatively quiet. Users can change it to "debug" for more verbose logging.