feat: Read dfx canister call argument from a file (#2323)
# Description
If the argument in `dfx canister call ... <ARGUMENT>` is large, e.g. 1.6Mb, the command fails because the command line is too long.  This is not due to a limit in dfx, that can be increased, but of operating systems such as Linux and OSX that make the command.  A solution is to allow providing the argument as a file, so that the command line is well within operating system bounds.

Fixes # (issue)

# How Has This Been Tested?

I have use the `--argument-file` flag to make API calls with large arguments.  An example:
```
# Some large payloads:
$ wc -c ,install-*
2687471 ,install-governance.arg
4453478 ,install-ledger.arg
1159432 ,install-root.arg
8300381 total
# Each is uploaded with:
$ /home/max/dfn/sdk/target/debug/dfx canister --network testnet call --argument-file ,install-root.arg nsrxe-iiaaa-aaaaa-aacaa-cai add_wasm 
(
  record {
    result = opt variant {
      Ok = record {
        hash = blob "\c9R\8a\1d\cc\27\c7\c2\a1e\8e\99\be\e8Z3\c4\c2-\a4\5c\84\81Cfkv\c0f=\5c\d6";
      }
    };
  },
)
```

> Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. Please also list any relevant details for your test configuration.

Note: If you direct me to where there are existing tests that the argument is read and transferred correctly I can add equivalent tests for `--argument-file`.

# Checklist:

- [x] The title of this PR complies with [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
- [x] I have edited the CHANGELOG accordingly.
- [x] I have made corresponding changes to the documentation.
  - I don't see any documentation affected by this change; if there is some, please direct me to it.