feat: Add additional logging from bitcoin canister in replica. (#2318)
# Description

Configures the replica to emit additional logging from the bitcoin
canister whenever the bitcoin feature is enabled. This helps show
useful information to developers, such as the bitcoin height that
the replica currently sees.

Fixes # EXC-1131

# How Has This Been Tested?

Ran it locally with a bitcoin enabled project and verified that I'm seeing
the logs from the bitcoin canister that I'm expecting to see.

# Checklist:

- [x] The title of this PR complies with [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
- [x] I have edited the CHANGELOG accordingly.
- [x] I have made corresponding changes to the documentation.