chore: add a comprehensive dfx json reference (#2361)
# Description

Adds `dfx schema`, a command that prints the current schema for `dfx.json`. Also added a workflow to keep a copy of the output up to date at `docs/dfx-json-schema.json` so that the docs can reference the current schema.

First part of [SDK-287](https://dfinity.atlassian.net/browse/SDK-287)

# How Has This Been Tested?

- manually
- workflow checks that `docs/dfx-json-schema.json` is up to date
- e2e checks that `dfx schema` produces valid json

# Checklist:

- [x] The title of this PR complies with [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
- [x] I have edited the CHANGELOG accordingly.
- [x] I have made corresponding changes to the documentation.