feat: check that release binaries are publishable, during PRs too (#2353)
# Description

The publish workflow checks that the release binaries dynamically link to a known set of libraries.  Sometimes it's fine to add more, but we want to find out before we ship.  This PR lets us find out earlier, when any PR introduces a new dependency.

This adds a new status, `publishable:required`, that we will add as a required status after merging this PR.

Fixes https://dfinity.atlassian.net/browse/SDK-593

# How Has This Been Tested?

Manually

- Build from PR: https://github.com/dfinity/sdk/runs/7366856852?check_suite_focus=true
- Build from tag: https://github.com/dfinity/sdk/runs/7366991008?check_suite_focus=true
  - https://github.com/dfinity/sdk/releases/tag/0.0.1-enstest.1