fix: improve JS language binding syntax compatibility (#2363)
# Description

This PR fixes a JavaScript syntax compatibility issue encountered in Motoko Playground (and more generally `create-react-app`) when compiling with Node v16 or v18 (dfinity/motoko-playground#116). When supporting a wider range of browser versions, the ES2020 optional chaining (?.) operator causes parsing issues which require configuring a Babel plugin for the generated `.dfx` directory. 

Since it's likely that other users will encounter this issue, this PR offers a more general fix in the form of replacing the (?.) operator with equivalent ternary expressions. 

# How Has This Been Tested?

Works as expected using an ES2016 interpreter. 

# Checklist:

- [x] The title of this PR complies with [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
- [x] I have edited the CHANGELOG accordingly.
- [x] I have made corresponding changes to the documentation (N/A).

@chenyan-dfinity