feat: remove internal webserver (#2390)
# Description

The only thing this was still serving was the `/_/candid` endpoint.  Since this served candid data for the current project, this was a blocker for the system-wide dfx feature.

Fixes https://dfinity.atlassian.net/browse/SDK-550

# How Has This Been Tested?

Removed relevant tests, and there are already tests for the `dfx canister metadata` command, which provides the same functionality.

# Checklist:

- [x] The title of this PR complies with [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
- [x] I have edited the CHANGELOG accordingly.
- [x] I have made corresponding changes to the documentation.